Fix concurrency bug
When using join to trigger an event, the consumer can complete before
the producer if the producer is not significantly faster than the
consumer. In this case, the queue will not output all inputs, and will
give no obvious indication that this is the case.

By marking the end of the list with None, we insure that both the
producer and consumer will process all inputs, and that both will return
appropriately. It is worth noting that this approach only works in
situations where the length of the input is known, as in this case.