Move to tsv format
Current gtex API is quite buggy. If hclust is specified at all (true or
false), it is run, forcing a json output. By not specifying hclust, we
can achieve either csv or tsv. But, when csv is specified, the first row
is tab-separated while the remainder is comma-separated.

Thus, we must 1) not specify hclust and 2) specify csv to get easily
readable output.