docs(index): add changelog
Adds link to changelog to documentation index. Currently, there is no
option to format the python semantic release Changelog as rst, and the
myst_parser for Sphinx (appears to) require the whole documentation to
be in Markdown. As such, a simple rst include directive doesn't work
here. Until one of the 2 above issues gets sorted, we default to linking
to the Github release history, which contains the whole changelog.